Vim: Conditionally initialize as vi or Vim
Configure Vim based on the command name.  If it is invoked as ex, vi, or
view, configure it to behave like a POSIX ex/vi.  If the running Vim was
compiled with only the "tiny" or "small" feature sets, treat it like vi
even if executed as vim.

If Vim is invoked as vim, gvim, or any of its other names, keep the
built-in and site initializations and load additional startup commands
from another file.